nomad_wod/requirements/development.txt
# coverage measures code coverage of Python programs
# See http://coverage.readthedocs.io/en/latest/
coverage

# **tox** automates testing
# See https://tox.readthedocs.io/en/latest/index.html
tox

# **Sphinx** is used to create documentation
# See http://www.sphinx-doc.org/en/stable/tutorial.html
Sphinx

# **flake8** is used to check for PEP8 compliance
# See http://flake8.pycqa.org/en/latest/index.html
flake8

social-auth-app-django
social-auth-app-mongoengine
django-crispy-forms
